:hammer: Move logic for binding desktop UI in AppWindowManager from commonMain to desktopMain
