+++ composeApp/src/desktopMain/kotlin/com/crosspaste/app/LinuxAppWindowManager.kt
@@ -16,7 +16,7 @@ class LinuxAppWindowManager(
     private val lazyShortcutKeys: Lazy<ShortcutKeys>,
     private val activeGraphicsDevice: ActiveGraphicsDevice,
     private val userDataPathProvider: UserDataPathProvider,
-) : AbstractAppWindowManager() {
+) : DesktopAppWindowManager() {
 
     private var prevLinuxAppInfo: LinuxAppInfo? by mutableStateOf(null)
