+++ composeApp/src/desktopMain/kotlin/com/crosspaste/ui/paste/detail/HtmlToImageDetailView.kt
