+++ composeApp/src/desktopMain/kotlin/com/crosspaste/ui/paste/detail/PasteDetailInfoView.kt
