+++ composeApp/src/desktopMain/kotlin/com/crosspaste/ui/search/DetialPasteDataView.kt
